refactor: Dao 인터페이스 추출해서 적용
