+++ src/main/java/chess/service/dto/BoardDto.java
@@ -5,19 +5,9 @@ import java.util.Map;
 
 public class BoardDto {
     private final List<PieceWithSquareDto> pieces;
-    private int id;
 
     public BoardDto(List<PieceWithSquareDto> pieces) {
         this.pieces = pieces;
-    }
-
-    public BoardDto(int id, List<PieceWithSquareDto> pieces) {
-        this.id = id;
-        this.pieces = pieces;
-    }
-
-    public int getId() {
-        return id;
     }
 
     public List<PieceWithSquareDto> getPieces() {
